refactor: 자주사용될만한 타입들 모아두기
