+++ src/api/roadmap/type.ts
@@ -1,3 +1,5 @@
+import type { Step, Category, Roadmap } from '@/api/type';
+
 // getRoadmaps
 export interface UserRoadmapsResponse {
   success: boolean;
@@ -8,36 +10,6 @@ export interface UserRoadmapsResponse {
 export interface UserRoadmapsResult {
   category: Category[];
   roadmap: Roadmap;
-}
-
-interface Category {
-  id: number;
-  name: string;
-}
-
-interface Roadmap {
-  tily: Tily[];
-  group: Group[];
-}
-
-interface Tily {
-  id: number;
-  name: string;
-  stepNum: number;
-}
-
-interface Group {
-  id: number;
-  name: string;
-  stepNum: number;
-  image: string;
-  creator: Creator;
-}
-
-interface Creator {
-  id: number;
-  name: string;
-  image: string;
 }
 
 // getRoadmapSteps
@@ -51,13 +23,6 @@ export interface RoadmapStepsResult {
   steps: Step[];
   progress: number;
   role: string;
-}
-
-export interface Step {
-  id: number;
-  title: string;
-  isCompleted: boolean;
-  tilId: number | null;
 }
 
 // postRoadmapsIndividual
