+++ src/api/user/type.ts
@@ -1,4 +1,4 @@
-import type { CommonResponse } from '@/api/type';
+import type { CommonResponse, UserHistory, User, Alarm } from '@/api/type';
 
 // getUserHistory
 export interface GetUserHistoryResponse {
@@ -9,21 +9,9 @@ export interface GetUserHistoryResponse {
   };
 }
 
-export interface UserHistory {
-  day: string;
-  value: number;
-}
-
 // getUser
 export interface GetUserResponse extends CommonResponse {
   result: User;
-}
-
-export interface User {
-  id: number;
-  name: string;
-  email: string;
-  image: string;
 }
 
 // getAlarm
@@ -31,30 +19,6 @@ export interface GetAlarmsResponse extends CommonResponse {
   result: {
     alarms: Alarm[];
   };
-}
-
-export interface Alarm {
-  id: number;
-  tilId: number;
-  isChecked: boolean;
-  roadmap: Roadmap;
-  step: Step;
-  sender: Sender;
-  createdAt: Date;
-}
-
-export interface Sender {
-  name: string;
-  image: string;
-}
-
-export interface Step {
-  id: number;
-  name: string;
-}
-export interface Roadmap {
-  id: number;
-  name: string;
 }
 
 // patchAlarm
